requirements.txt
langchain
pinecone-client
python-dotenv
google-genai
google-auth
google-auth-oauthlib
google-cloud-aiplatform
